backend/requirements.txt
# Framework principal
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Base de datos
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Validación y serialización
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12

# Tareas asíncronas
celery==5.4.0
redis==5.1.1

# HTTP client (conectores GPS, webhooks)
httpx==0.27.2

# Notificaciones
firebase-admin==6.5.0
sendgrid==6.11.0

# Utilidades
python-dotenv==1.0.1
pytz==2024.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Linting y formato
ruff==0.6.9
